july25/mcp-socialmedia/pyproject.toml
[project]
name = "mcp-socialmedia"
version = "0.1.0"
description = "The Model Context protocol for social media postings"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "linkedin-api-client>=0.3.0",
    "mcp[cli]>=1.12.0",
    "requests>=2.32.4",
]

